+++ src/imgw_data/export/export.py
@@ -20,7 +20,7 @@ def _check_filesuffix(fname: str, suffix: str):
     return fname + '.' + suffix
 
 
-def export_to_file(ds, fname: str, as_json=False, as_csv=False, as_xml=False, as_html=False):
+def export_to_file(ds, fname: str, ftype: str):
     """
     Function exports dataset to a specific file.
 
@@ -30,16 +30,10 @@ def export_to_file(ds, fname: str, as_json=False, as_csv=False, as_xml=False, as
 
     fname : str
 
-    as_json : bool, default=False
-
-    as_csv : bool, default=False
-
-    as_xml : bool, default=False
-
-    as_html : bool, default=False
+    ftype : str
     """
 
-    if as_json:
+    if ftype == 'json':
         fname = _check_filesuffix(fname, suffix='json')
         if isinstance(ds, list):
             with open(fname, 'w', encoding='utf8') as fin:
@@ -49,14 +43,14 @@ def export_to_file(ds, fname: str, as_json=False, as_csv=False, as_xml=False, as
                 serialized = json.dumps(ds, ensure_ascii=False).encode('utf-8')
                 fin.write(serialized)
     else:
-        if as_csv:
+        if ftype == 'csv':
             fname = _check_filesuffix(fname, suffix='csv')
-        elif as_xml:
+        elif ftype == 'xml':
             fname = _check_filesuffix(fname, suffix='xml')
-        elif as_html:
+        elif ftype == 'html':
             fname = _check_filesuffix(fname, suffix='html')
         else:
-            raise TypeError('Output file format must be set!')
+            raise TypeError('Unknown output file type!')
 
         with open(fname, 'w', encoding='utf8') as fin:
             fin.write(ds)