+++ src/imgw_data/utils/parse.py
@@ -3,7 +3,56 @@ from typing import Union
 import requests
 
 
-def parse_response(response: requests.Response, baseurl: str) -> Union[list[dict], str]:
+def add_coordinates(readings: list[dict],
+                    coordinates: list[list],
+                    raise_error_if_missing=False):
+    """
+    Function adds coordinates to downloaded stations.
+
+    Parameters
+    ----------
+    readings : list[dict]
+        Current weather readings.
+
+    coordinates : list[list]
+        ``[[id, longitude, latitude], ...]``
+
+    raise_error_if_missing : bool, default = False
+        If any coordinates are missing then raise error.
+
+    Returns
+    -------
+    : list[dict]
+        Readings woth additional keys: ``lon`` | ``lat``.
+    """
+
+    transformed = []
+    transformed_ids = set()
+    all_ids = set()
+
+    for reading in readings:
+        rid = reading['station_id'][2:]
+        all_ids.add(rid)
+        # Get coordinates
+        for coo in coordinates:
+            if str(coo[0]) == rid:
+                new_reading = reading.copy()
+                new_reading['lon'] = coo[1]
+                new_reading['lat'] = coo[2]
+                transformed.append(new_reading)
+                transformed_ids.add(str(coo[0]))
+                break
+
+    if raise_error_if_missing:
+        if len(transformed) != len(readings):
+            missing_ids = all_ids.difference(transformed_ids)
+            missing_stations = ['12'+str(x) for x in missing_ids]
+            raise ValueError(f'Missing coordinates for those stations: {missing_stations}')
+
+    return transformed
+
+
+def parse_response(response: requests.Response, ftype: str) -> Union[list[dict], str]:
     """
     Function parses response data to appropriate format.
 
@@ -11,17 +60,17 @@ def parse_response(response: requests.Response, baseurl: str) -> Union[list[dict
     ----------
     response : requests.Response
 
-    baseurl : str
+    ftype : str
 
     Returns
     -------
     : Union[list[dict], str]
     """
-    if baseurl.endswith('csv'):
+    if ftype == 'csv':
         return response.text
-    elif baseurl.endswith('xml'):
+    elif ftype == 'xml':
         return response.text
-    elif baseurl.endswith('html'):
+    elif ftype == 'html':
         return response.text
 
     return response.json()