+++ tests/test_load.py
@@ -1,0 +1,6 @@
+from imgw_data import get_current_weather
+
+
+def test_load_data():
+    weather = get_current_weather()
+    assert bool(weather)