+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+## Version 0.1.2 (2024-04-14)
+
+- (bug) circular import is removed
+- (data) Added coordinates to the JSON output
+
 ## Version 0.1.1 (2024-04-12)
 
 - initial version of the package with basic functionalities, downloads current weather data